requirements.txt
django==4.1.2
djangorestframework==3.14
flake8==6.0.0
black==23.7.0